LPS-72240 Pull out NullClassTypeReader as a standalone class
