+++ portal-kernel/src/com/liferay/asset/kernel/model/BaseAssetRendererFactory.java
@@ -14,7 +14,6 @@
 
 package com.liferay.asset.kernel.model;
 
-import com.liferay.asset.kernel.NoSuchClassTypeException;
 import com.liferay.asset.kernel.service.AssetEntryLocalServiceUtil;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.language.LanguageUtil;
@@ -36,7 +35,6 @@ import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.util.WebKeys;
 
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Locale;
@@ -425,23 +423,5 @@ public abstract class BaseAssetRendererFactory<T>
 	private boolean _searchable;
 	private boolean _selectable = true;
 	private boolean _supportsClassTypes;
-
-	private static class NullClassTypeReader implements ClassTypeReader {
-
-		@Override
-		public List<ClassType> getAvailableClassTypes(
-			long[] groupIds, Locale locale) {
-
-			return Collections.emptyList();
-		}
-
-		@Override
-		public ClassType getClassType(long classTypeId, Locale locale)
-			throws PortalException {
-
-			throw new NoSuchClassTypeException();
-		}
-
-	}
 
 }
